Added a CAN-blink test.
Designed to run on two teensies: A receiver and a transmitter. The transmitter sends alternatng 1 byte CAN messages of 0 and 1, and the receiver listens for these messages and sets its LED according to the state of the message. One thing to note is that this receiver code works with the hello_world example built by Josh Christensen (me) and Bjorn Neergard, and that transmitter code works with this receiver code, but the hello_world receiver code doesn't seem to work with this transmitter code. Will look into when I have a logic analyzer or a scope.